+++ sdk/cosmos/azure-cosmos/tests/test_fault_injection_transport.py
@@ -297,6 +297,7 @@ class TestFaultInjectionTransport:
 
         initialized_objects = self.setup_method_with_custom_transport(
             custom_transport,
+            multiple_write_locations=True,
             preferred_locations=["First Region", "Second Region"])
         container: ContainerProxy = initialized_objects["col"]
 
@@ -347,6 +348,7 @@ class TestFaultInjectionTransport:
 
         initialized_objects = self.setup_method_with_custom_transport(
             custom_transport,
+            multiple_write_locations=True,
             preferred_locations=["First Region", "Second Region"])
         container: ContainerProxy = initialized_objects["col"]
 
@@ -466,7 +468,9 @@ class TestFaultInjectionTransport:
 
         initialized_objects = self.setup_method_with_custom_transport(
             custom_transport,
-            preferred_locations=["First Region", "Second Region"])
+            preferred_locations=["First Region", "Second Region"],
+            multiple_write_locations=True
+        )
         container: ContainerProxy = initialized_objects["col"]
         with pytest.raises(ServiceRequestError):
             container.upsert_item(body=document_definition)