+++ sdk/cosmos/azure-cosmos/tests/test_location_cache.py
@@ -62,17 +62,17 @@ class TestLocationCache:
 
     def test_is_endpoint_unavailable(self):
         lc = refresh_location_cache([], False)
-        assert lc.is_endpoint_unavailable_internal(location1_endpoint, "Read") is False
-        assert lc.is_endpoint_unavailable_internal(location1_endpoint, "None") is False
-        assert lc.is_endpoint_unavailable_internal(location1_endpoint, "Write") is False
+        assert lc.is_endpoint_unavailable(location1_endpoint, "Read") is False
+        assert lc.is_endpoint_unavailable(location1_endpoint, "None") is False
+        assert lc.is_endpoint_unavailable(location1_endpoint, "Write") is False
         lc.mark_endpoint_unavailable_for_read(location1_endpoint, False)
-        assert lc.is_endpoint_unavailable_internal(location1_endpoint, "Read")
-        assert lc.is_endpoint_unavailable_internal(location1_endpoint, "None") is False
-        assert lc.is_endpoint_unavailable_internal(location1_endpoint, "Write") is False
+        assert lc.is_endpoint_unavailable(location1_endpoint, "Read")
+        assert lc.is_endpoint_unavailable(location1_endpoint, "None") is False
+        assert lc.is_endpoint_unavailable(location1_endpoint, "Write") is False
         lc.mark_endpoint_unavailable_for_write(location2_endpoint, False)
-        assert lc.is_endpoint_unavailable_internal(location2_endpoint, "Read") is False
-        assert lc.is_endpoint_unavailable_internal(location2_endpoint, "None") is False
-        assert lc.is_endpoint_unavailable_internal(location2_endpoint, "Write")
+        assert lc.is_endpoint_unavailable(location2_endpoint, "Read") is False
+        assert lc.is_endpoint_unavailable(location2_endpoint, "None") is False
+        assert lc.is_endpoint_unavailable(location2_endpoint, "Write")
         location1_info = lc.location_unavailability_info_by_endpoint[location1_endpoint]
         lc.location_unavailability_info_by_endpoint[location1_endpoint] = location1_info
 
@@ -83,8 +83,7 @@ class TestLocationCache:
 
         # check endpoints to health check
         endpoints = lc.endpoints_to_health_check()
-        assert len(endpoints) == 3
-        assert default_endpoint in endpoints
+        assert len(endpoints) == 2
         assert location1_endpoint in endpoints
         assert location4_endpoint in endpoints
 
@@ -107,7 +106,7 @@ class TestLocationCache:
             found_endpoint = False
             endpoint = read_region['databaseAccountEndpoint']
             for region in read_regions:
-                if endpoint in (region.get_primary(), region.get_alternate()):
+                if endpoint == region.get_primary():
                     found_endpoint = True
             assert found_endpoint
 
@@ -118,7 +117,7 @@ class TestLocationCache:
             found_endpoint = False
             endpoint = write_region['databaseAccountEndpoint']
             for region in write_regions:
-                if endpoint in (region.get_primary(), region.get_alternate()):
+                if endpoint == region.get_primary():
                     found_endpoint = True
             assert found_endpoint