+++ sdk/cosmos/azure-cosmos/tests/test_regional_routing_context.py
@@ -17,7 +17,7 @@ class TestRegionalRoutingContext(unittest.TestCase):
     REGION1 = "West US"
     REGION2 = "East US"
     REGION3 = "West US 2"
-    REGIONAL_ROUTING_CONTEXT = RegionalRoutingContext(host, "something_different")
+    REGIONAL_ROUTING_CONTEXT = RegionalRoutingContext(host)
     TEST_DATABASE_ID = test_config.TestConfig.TEST_DATABASE_ID
     TEST_CONTAINER_ID = test_config.TestConfig.TEST_SINGLE_PARTITION_CONTAINER_ID